fix: robust reconnect logic for AI workflows and cache invalidation (Issue #54)
